[Performance] Don't add `org-journal-redraw-calendar` to `after-save-hook` and `after-revert-hook`
When the function `org-journal-redraw-calendar` gets called the function
`org-journal-mark-entries` gets called two times, plus it is unnecessary
because the calendar may not even be visible.

One time because the function `org-journal-redraw-calendar` is called by the
`after-save-hook` and one time when the calendar is opened inside
`org-journal-redraw-calendar`, that is because opening the calendar triggers
the hook `calendar-today-visible-hook`.

Bind key 'm' to `org-journal-mark-entries` inside calendar.