Allow changing of weekly journal start day.
Added two customization options org-journal-start-on-weekday
and org-journal-use-agenda-start-on-weekday.  The first changes
the start day of daily journals.  The second tells org-journal
to use the value of org-agenda-start-on-weekday for its start
day.  When the second option is set, the first is ignored.

Do note that established journals will no longer work when the
start day is changed.