Solve error when org-journal-dir has symlink in the path
Solve the error "Failed to extract year from file: %s" for journal file with
link in their path.

Inside function org-journal--file-name->calendar-date there is a mismatch
between ‘file-pattern’ and the parameter ‘file-name’. Because file-pattern has
its path with symlink resolved with file-truename but the parameter file-name
still have the path with symlink.