Update org-journal.el
Set major-mode for org-journal pop-up buffer to org-mode

Credit to @zhyzky for pinpointing the issue and describing the fix in #414. There's a slightly different macro described in the issue report. However I wasn't able to get it working so I opted instead to just take the solution and modify the existing code as minimally as possible.