Allow any alpha character in date format
Some locales have day names with accents, such as Esperanto.  When
using org-journal in such a locale and customizing
`org-journal-created-property-timestamp-format` to support org
inactive timestamps, as `[%Y-%m-%d %a]` org-journal will fail with an
error on many operations:

    org-journal--entry-date->calendar-date: Created property timestamp format "[%Y-%m-%d %a]" doesn’t match CREATED property value ([2021-12-02 ĵaŭ]) from entry at line: Line 29

This is because the `%a` format specifier in org-journal only accepts
unaccented day names.  This patch allows `%a` to accept any character
Emacs classifies as alpha.